+++ src/main/java/org/fms/eis/webapp/action/PDaserverGroupAction.java
@@ -9,6 +9,8 @@ package org.fms.eis.webapp.action;
 import com.riozenc.titanTool.spring.web.http.HttpResult;
 import com.riozenc.titanTool.spring.web.http.HttpResultPagination;
 import org.fms.eis.webapp.service.IPDaserverGroupService;
+import org.fms.eis.webapp.service.IPSysNodeService;
+import org.fms.eis.webapp.vo.PChnlGroupVO;
 import org.fms.eis.webapp.vo.PDaserverGroupStaticVO;
 import org.fms.eis.webapp.vo.PDaserverGroupVO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,6 +30,10 @@ public class PDaserverGroupAction {
     @Autowired
     @Qualifier("PDaserverGroupServiceImpl")
     private IPDaserverGroupService pDaserverGroupService;
+
+    @Autowired
+    @Qualifier("PSysNodeServiceImpl")
+    private IPSysNodeService pSysNodeService;
 
     @ResponseBody
     @PostMapping(params = "method=insert")
@@ -55,11 +61,32 @@ public class PDaserverGroupAction {
 
     }
 
+    /**
+     * 已关联采集主机的不能删除
+     *
+     * @return
+     */
     @ResponseBody
     @PostMapping(params = "method=delete")
     public HttpResult<?> delete(@RequestBody List<PDaserverGroupVO> deleteList) throws Exception {
-        HttpResult httpResult = pDaserverGroupService.deleteList(deleteList);
-        return httpResult;
+        if (deleteList != null) {
+            if (deleteList.size() > 0) {
+                String value = "";
+                for (PDaserverGroupVO item : deleteList) {
+                    value += item.getId() + ",";
+                }
+                value = value.substring(0, value.length() - 1);
+                if (pSysNodeService.findByRelDasGroup(value).size() > 0) {
+                    return new HttpResult<String>(HttpResult.ERROR, "已关联采集主机的不能删除", null);
+                } else {
+                    return pDaserverGroupService.deleteList(deleteList);
+                }
+            } else {
+                return new HttpResult<String>(HttpResult.ERROR, "暂无要删除的内容", null);
+            }
+        } else {
+            return new HttpResult<String>(HttpResult.ERROR, "参数传递错误", null);
+        }
     }
 
     @ResponseBody
